+++ config/mail/email_handle.js
@@ -1,0 +1,14 @@
+const API_KEY = 'YOUR_API_KEY';
+const DOMAIN = 'YOUR_DOMAIN_NAME';
+const email_handle = require('mailgun-js')({apiKey: process.env.MAILGUN_API_KEY, domain: process.env.MAILGUN_DOMAIN});
+
+const data = {
+  from: 'Excited User <me@samples.email_handle.org>',
+  to: 'foo@example.com, bar@example.com',
+  subject: 'Hello',
+  text: 'Testing some Mailgun awesomeness!'
+};
+
+email_handle.messages().send(data, (error, body) => {
+  console.log(body);
+});
